Correct mbed timestamp bug

+++ libraries/mbed/common/us_ticker_api.c
@@ -37,7 +37,9 @@ void us_ticker_irq_handler(void) {
             return;
         }
 
-        if ((int64_t)(head->timestamp - us_ticker_read()) <= 0) {
+        // if ((int64_t)(head->timestamp - us_ticker_read()) <= 0) 
+        if ((int32_t)((uint32_t)head->timestamp - us_ticker_read()) <= 0) // betzw: just 32-bits available
+	{
             // This event was in the past:
             //      point to the following one and execute its handler
             ticker_event_t *p = head;
